[AIP-21] Move contrib.hooks.redis_hook providers.redis.hooks.redis
