+++ airflow/contrib/hooks/redis_hook.py
@@ -16,53 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""This module is deprecated. Please use `airflow.providers.redis.hooks.redis`."""
 
-"""
-RedisHook module
-"""
-from redis import Redis
+import warnings
 
-from airflow.hooks.base_hook import BaseHook
+# pylint: disable=unused-import
+from airflow.providers.redis.hooks.redis import RedisHook  # noqa
 
-
-# noinspection PyAbstractClass
-class RedisHook(BaseHook):
-    """
-    Wrapper for connection to interact with Redis in-memory data structure store
-    """
-    def __init__(self, redis_conn_id='redis_default'):
-        """
-        Prepares hook to connect to a Redis database.
-
-        :param conn_id:     the name of the connection that has the parameters
-                            we need to connect to Redis.
-        """
-        self.redis_conn_id = redis_conn_id
-        self.redis = None
-        self.host = None
-        self.port = None
-        self.password = None
-        self.db = None
-
-    def get_conn(self):
-        """
-        Returns a Redis connection.
-        """
-        conn = self.get_connection(self.redis_conn_id)
-        self.host = conn.host
-        self.port = conn.port
-        self.password = None if str(conn.password).lower() in ['none', 'false', ''] else conn.password
-        self.db = conn.extra_dejson.get('db', None)
-
-        if not self.redis:
-            self.log.debug(
-                'Initializing redis object for conn_id "%s" on %s:%s:%s',
-                self.redis_conn_id, self.host, self.port, self.db
-            )
-            self.redis = Redis(
-                host=self.host,
-                port=self.port,
-                password=self.password,
-                db=self.db)
-
-        return self.redis
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.redis.hooks.redis`.",
+    DeprecationWarning, stacklevel=2
+)
