+++ tests/contrib/operators/test_redis_publish_operator.py
@@ -24,8 +24,8 @@ import pytest
 from mock import MagicMock
 
 from airflow import DAG
-from airflow.contrib.hooks.redis_hook import RedisHook
 from airflow.contrib.operators.redis_publish_operator import RedisPublishOperator
+from airflow.providers.redis.hooks.redis import RedisHook
 from airflow.utils import timezone
 
 DEFAULT_DATE = timezone.datetime(2017, 1, 1)
