+++ tests/contrib/sensors/test_redis_sensor.py
@@ -23,8 +23,8 @@ import unittest
 import pytest
 
 from airflow import DAG
-from airflow.contrib.hooks.redis_hook import RedisHook
 from airflow.contrib.sensors.redis_key_sensor import RedisKeySensor
+from airflow.providers.redis.hooks.redis import RedisHook
 from airflow.utils import timezone
 
 DEFAULT_DATE = timezone.datetime(2017, 1, 1)
